finished color items

+++ MarvelSandbox/src/Core/PythonInterfaces/mvDrawingInterface.h
@@ -25,14 +25,14 @@ namespace Marvel{
 				{mvPythonDataType::FloatList, "pmax"},
 				{mvPythonDataType::FloatList, "uv_min"},
 				{mvPythonDataType::FloatList, "uv_max"},
-				{mvPythonDataType::FloatList, "color"}
+				{mvPythonDataType::IntList, "color"}
 			}, false, "Needs documentation")},
 
 			{"drawLine", mvPythonTranslator({
 				{mvPythonDataType::String, "drawing"},
 				{mvPythonDataType::FloatList, "p1"},
 				{mvPythonDataType::FloatList, "p2"},
-				{mvPythonDataType::FloatList, "color"},
+				{mvPythonDataType::IntList, "color"},
 				{mvPythonDataType::Integer, "thickness"},
 			}, false, "Needs documentation")},
 
@@ -41,7 +41,7 @@ namespace Marvel{
 				{mvPythonDataType::FloatList, "p1"},
 				{mvPythonDataType::FloatList, "p2"},
 				{mvPythonDataType::FloatList, "p3"},
-				{mvPythonDataType::FloatList, "color"},
+				{mvPythonDataType::IntList, "color"},
 				{mvPythonDataType::Optional},
 				{mvPythonDataType::FloatList, "fill"},
 				{mvPythonDataType::Float, "thickness"}
@@ -51,7 +51,7 @@ namespace Marvel{
 				{mvPythonDataType::String, "drawing"},
 				{mvPythonDataType::FloatList, "pmin"},
 				{mvPythonDataType::FloatList, "pmax"},
-				{mvPythonDataType::FloatList, "color"},
+				{mvPythonDataType::IntList, "color"},
 				{mvPythonDataType::Optional},
 				{mvPythonDataType::FloatList, "fill"},
 				{mvPythonDataType::Float, "rounding"},
@@ -64,7 +64,7 @@ namespace Marvel{
 				{mvPythonDataType::FloatList, "p2"},
 				{mvPythonDataType::FloatList, "p3"},
 				{mvPythonDataType::FloatList, "p4"},
-				{mvPythonDataType::FloatList, "color"},
+				{mvPythonDataType::IntList, "color"},
 				{mvPythonDataType::Optional},
 				{mvPythonDataType::FloatList, "fill"},
 				{mvPythonDataType::Float, "thickness"}
@@ -75,7 +75,7 @@ namespace Marvel{
 				{mvPythonDataType::FloatList, "pos"},
 				{mvPythonDataType::String, "text"},
 				{mvPythonDataType::Optional},
-				{mvPythonDataType::FloatList, "color"},
+				{mvPythonDataType::IntList, "color"},
 				{mvPythonDataType::Integer, "size"}
 			}, false, "Needs documentation")},
 
@@ -83,7 +83,7 @@ namespace Marvel{
 				{mvPythonDataType::String, "drawing"},
 				{mvPythonDataType::FloatList, "center"},
 				{mvPythonDataType::Float, "radius"},
-				{mvPythonDataType::FloatList, "color"},
+				{mvPythonDataType::IntList, "color"},
 				{mvPythonDataType::Optional},
 				{mvPythonDataType::Integer, "segments"},
 				{mvPythonDataType::Float, "thickness"},
@@ -93,7 +93,7 @@ namespace Marvel{
 			{"drawPolyline", mvPythonTranslator({
 				{mvPythonDataType::String, "drawing"},
 				{mvPythonDataType::FloatList, "points"},
-				{mvPythonDataType::FloatList, "color"},
+				{mvPythonDataType::IntList, "color"},
 				{mvPythonDataType::Optional},
 				{mvPythonDataType::Integer, "closed"},
 				{mvPythonDataType::Float, "thickness"}
@@ -102,7 +102,7 @@ namespace Marvel{
 			{"drawPolygon", mvPythonTranslator({
 				{mvPythonDataType::String, "drawing"},
 				{mvPythonDataType::FloatList, "points"},
-				{mvPythonDataType::FloatList, "color"},
+				{mvPythonDataType::IntList, "color"},
 				{mvPythonDataType::Optional},
 				{mvPythonDataType::FloatList, "fill"},
 				{mvPythonDataType::Float, "thickness"}
@@ -114,7 +114,7 @@ namespace Marvel{
 				{mvPythonDataType::FloatList, "p2"},
 				{mvPythonDataType::FloatList, "p3"},
 				{mvPythonDataType::FloatList, "p4"},
-				{mvPythonDataType::FloatList, "color"},
+				{mvPythonDataType::IntList, "color"},
 				{mvPythonDataType::Optional},
 				{mvPythonDataType::Float, "thickness"},
 				{mvPythonDataType::Integer, "segments"}
